requirements.txt
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
futures==3.2.0
google-api-core==1.0.0
google-auth==1.4.1
google-cloud-core==0.28.1
google-cloud-translate==1.3.1
googleapis-common-protos==1.5.3
idna==2.6
protobuf==3.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pytz==2018.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
SpeechRecognition==3.8.1
urllib3==1.22
